Fix: 修复上一次提交的问题.

+++ src/model.rs
@@ -9,16 +9,16 @@ pub mod weather;
 pub const CAMPUS_FENGXIAN: i32 = 1;
 pub const CAMPUS_XUHUI: i32 = 2;
 
-const DEFAULT_PAGE_INDEX: u16 = 0;
-const DEFAULT_ITEM_COUNT: u16 = 20;
+const DEFAULT_PAGE_INDEX: i32 = 0;
+const DEFAULT_ITEM_COUNT: i32 = 20;
 
 /// Page parameters for list pagination
-#[derive(Serialize, Deserialize, Default)]
+#[derive(serde::Serialize, serde::Deserialize, Default)]
 pub struct PageView {
     /// Page index, 1 is the minimum value
-    pub index: Option<u16>,
+    pub index: Option<i32>,
     /// Page count, 1 is the minimum value
-    pub count: Option<u16>,
+    pub count: Option<i32>,
 }
 
 impl PageView {
@@ -27,7 +27,7 @@ impl PageView {
         PageView::default()
     }
     /// Get validated index
-    pub fn index(&self) -> u16 {
+    pub fn index(&self) -> i32 {
         if let Some(index) = self.index {
             if index > 0 {
                 return index;
@@ -36,7 +36,7 @@ impl PageView {
         DEFAULT_PAGE_INDEX
     }
     /// Get validated item count value
-    pub fn count(&self, max_count: u16) -> u16 {
+    pub fn count(&self, max_count: i32) -> i32 {
         if let Some(count) = self.count {
             if count < max_count {
                 return count;
@@ -45,7 +45,7 @@ impl PageView {
         DEFAULT_ITEM_COUNT
     }
     /// Calculate offset
-    pub fn offset(&self, max_count: u16) -> u16 {
+    pub fn offset(&self, max_count: i32) -> i32 {
         self.count(max_count) * (self.index() - 1)
     }
 }
